fix: build error 수정, square uncrop 페이지 로직 동일
